code: merge all files into appli.lua
It's really not convenient to have all the files separated as it is
difficult to cross-reference them and it also forces to export more
functions than really needed. They were initially split so as to
limit the amount of memory required to compile each of them and to
avoid keeping most of them loaded at the same time (e.g. the screen
handlers). This is not a problem anymore with LFS. So Let's merge
them all into appli.lua and make most variables and functions local.