code: make it easier to set/unset the alarm
Instead of losing all digits past 23:59 when adjusting the alarm time,
let's make the hour's first digit cycle through 0,10,20,-10 and keep
the rest unchanged. This allows to adjust the first digit later to
easily re-enable the alarm time. Value "-10" is printed as "-0", so
that we always have exactly "-" instead of the tens digit when the
alarm is disabled.