doc: update the doc for LFS usage
The code is way smaller now with LFS, and leaves 28 kB of free memory
once everything is enabled. This also eases configuration changes and
allows to reconfigure the local time without dying on out-of-memory
conditions. The procedure is a bit more complex as it requires to build
the firmware image with a few new options, though nodemcu-build.com is
still usable for this.