code: update alarm digits from right to left
It was cumbersome to do it left-to-right since missing a digit on
units increases tens and requires a full-round of digits to update
tens again. Now we can start right till the desired digit is found,
then a single press swiches to tens, allowing to update or fix the
effect of previous digit update.