code: only switch from idle to full_start during the first 2 seconds
It's currently impossible to stop the raising light using the buttons
because setting the state back to LS_IDLE causes it to be switched to
LS_FULL_START in tick() as long as the hour and minute match. Let's also
check the second and only allow that transition to happen at seconds 0&1.